+++ src/signal.h
@@ -55,7 +55,7 @@ static Signal signals[] = {
 	/* signum           function */
 	{ "center",         center },
 	{ "cyclelayout",    cyclelayout },
-	{ "expose",			exposeview },
+	{ "winview",		winview },
 	{ "focusmon",       focusmon },
 	{ "focusstack",     focusstack },
 	{ "incnmaster",     incnmaster },